Reccomand/recsys/chat/requirements.txt
openai>=1.0.0
langchain>=0.0.300
faiss-cpu>=1.7.4
numpy>=1.24.0
cryptography>=41.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
aiomysql
acquire